* more device host_image functionality:    * added initial (bi/tri)linear sampling support (this is still wip and only enabled for 1D and 2D images right now, and still producing artifacts)    * added fp lod and bias support (concluding lod support), these act identical to the integer lod (until such time where a lod level is computed automatically, which will only happen in fragments shaders)    * rearranged the whole clamp dim and mip-level info data: this is now a neatly aligned 64 bytes per mip-level chunk of data (still 16 levels in total), which contains the per-level image dimension, integer clamp dimension (dim - 1), float clamp dimension (image dim converted to float) and level offset in bytes    * fixed the fp nearest coordinate computation/conversion: the normalized coord should be scaled with image dim (not dim-1), the +0.5f is unnecessary and the final value should be clamped to [0, image dim - eps] (not [0, image dim - 1]), this is less trivial than it looks like ...    * added cube map read/write with offset after all (while this is not supported by other backends, I do need this myself when doing linear sampling), note that the offset is an int2 as it is applied after computing the cube face (i.e. it's an offset in the 2D face image)    * I noticed that __builtin_choose_expr isn't actually evaluated at compile-time when the condition is type/value dependent (can't really do anything about this though, will have to wait until constexpr/static-if gets into the standard; use lto, this seems to help, dce doesn't for whatever reason)  * const_math and vector lib: added interpolate/cubic_interpolate/catmull_rom_interpolate for non-floating-point types (interpolator is still fp), let's see how will this goes  * vector lib: added fractional/fractionaled functions, these return a vector with the components set to the fractional part of the components of the input vector ("modf")

+++ compute/device/host_limits.hpp
@@ -24,14 +24,14 @@ namespace host_limits {
 	static constexpr const size_t local_memory_size { 128ull * 1024ull * 1024ull };
 	
 	//! max supported image dim, identical for all image types
-	static constexpr const size_t max_image_dim { 32768 };
+	static constexpr const uint32_t max_image_dim { 32768 };
 	
 	//! max amount of mip-map/lod levels that can exist (log2(max_image_dim) + 1)
-	static constexpr const size_t max_mip_levels { 16 };
+	static constexpr const uint32_t max_mip_levels { 16 };
 	static_assert((1u << (max_mip_levels - 1u)) == max_image_dim, "max #lod-levels must match max image dim");
 	
 	//! max work-items per work-group (max #fibers)
-	static constexpr const size_t max_work_group_size {
+	static constexpr const uint32_t max_work_group_size {
 #if !defined(__WINDOWS__)
 		1024
 #else // due to memory restrictions with windows fibers, this shouldn't be higher than 64
